+++ src/extensions/techno/technoext_hooks.cpp
@@ -44,6 +44,7 @@
 #include "fetchres.h"
 #include "hooker.h"
 #include "house.h"
+#include "houseext.h"
 #include "housetype.h"
 #include "infantry.h"
 #include "infantrytype.h"
@@ -127,6 +128,8 @@ public:
     int _Refund_Amount() const;
     bool _Evaluate_Object(ThreatType method, int mask, int range, TechnoClass const* object, int& value, int zone, Coord const& coord) const;
     bool _Should_Self_Heal_Now() const;
+    int _Apparent_Brightness(int brightness) const;
+    void _Flashing_AI();
 };
 
 
@@ -2212,28 +2215,16 @@ DEFINE_HOOK(0x0062C5D5, _TechnoClass_Draw_Health_Bars_Unit_Draw_Pos_Patch, 0)
 }
 
 
-/**
- *  #issue-411
- * 
- *  Implements IsAffectsAllies for WarheadTypes.
- * 
- *  @note: This patch does not replace "stolen" code as per our implementation
- *         rules, this is because the call to ObjectClass::Take_Damage that follows
- *         is too much of a risk to not have correctly implemented.
- * 
- *  @author: CCHyper
- */
-DEFINE_HOOK(0x006328DE, _TechnoClass_Take_Damage_IsAffectsAllies_Patch, 7)
+static bool Should_Take_Damage(TechnoClass* this_ptr, TechnoClass* source, const WarheadTypeClass* warhead, int damage)
 {
-    GET(TechnoClass *, this_ptr, ESI);
-    GET_STACK(int *, damage, 0xEC);
-    GET_STACK(const WarheadTypeClass *, warhead, 0xF4);
-    GET_STACK(TechnoClass *, source, 0xF8);
-
     if (warhead) {
 
         /**
+         *  #issue-411
+         * 
          *  Is the warhead that hit us one that affects units allied with its firing owner?
+         *
+         *  @author: CCHyper
          */
         WarheadTypeClassExtension* warheadtypeext = Extension::Fetch(warhead);
         if (!warheadtypeext->IsAffectsAllies) {
@@ -2243,12 +2234,50 @@ DEFINE_HOOK(0x006328DE, _TechnoClass_Take_Damage_IsAffectsAllies_Patch, 7)
              *  the damage amount and return that we took no damage.
              */
             if (source && source->House->Is_Ally(this_ptr->House)) {
-                *damage = 0;
-                goto return_RESULT_NONE;
+                return false;
             }
-
         }
+    }
 
+    /**
+     *  #issue-425
+     *
+     *  If this unit is shielded by the Iron Curtain, no damage should be dealt.
+     *
+     *  Author: Rampastring
+     */
+    TechnoClassExtension* technoext = Extension::Fetch(this_ptr);
+    if (technoext->IronCurtainTimer > 0) {
+        return false;
+    }
+
+    /**
+     *  If this unit is a priority for shielding with the Iron Curtain and it would take major damage, shield it.
+     */
+    TechnoTypeClassExtension* technotypeext = Extension::Fetch(this_ptr->TClass);
+    if (technotypeext->IronCurtainPriorityTarget && this_ptr->Strength - damage <= Rule->ConditionYellow * this_ptr->TClass->MaxStrength) {
+        if (technoext->Iron_Curtain_Me(false)) {
+            Extension::Fetch(this_ptr->House)->Expend_Iron_Curtain();
+            return false;
+        }
+    }
+
+    return true;
+}
+
+/**
+ *  Implements IsAffectsAllies for WarheadTypes as well as the Iron Curtain effect for Technos.
+ */
+DEFINE_HOOK(0x006328DE, _TechnoClass_Take_Damage_Intercept_Patch, 7)
+{
+    GET(TechnoClass*, this_ptr, ESI);
+    GET_STACK(int*, damage, 0xEC);
+    GET_STACK(const WarheadTypeClass*, warhead, 0xF4);
+    GET_STACK(TechnoClass*, source, 0xF8);
+
+    if (!Should_Take_Damage(this_ptr, source, warhead, *damage)) {
+        *damage = 0;
+        goto return_RESULT_NONE;
     }
 
     return 0;
@@ -3187,6 +3216,66 @@ DEFINE_HOOK(0x0062AB5E, _TechnoClass_Revealed_Look_For_Allies_Patch, 0)
 
 
 /**
+ *  Implements flashing for the Iron Curtain.
+ *
+ *  Author: tomsons26/ZivDero for original function code, Rampastring for Iron Curtain effect.
+ */
+int TechnoClassExt::_Apparent_Brightness(int brightness) const
+{
+    TechnoClassExtension* technoext = Extension::Fetch(this);
+
+    if (technoext->IronCurtainTimer > 0) {
+        int index = (technoext->IronCurtainTimer.Value() / RuleExtension->IronCurtainFlashRate) % RuleExtension->IronCurtainPulseTable.Count();
+        int extra = RuleExtension->IronCurtainPulseTable[index] * RuleExtension->IronCurtainFlashIntensityMultiplier;
+        int total = brightness + extra;
+        if (total < 0) total = 0;
+
+        return total;
+    }
+
+    if (((FlashCount / 2) % 2) == 1) {
+        return (brightness > 1500 ? 500 : 2000);
+    } else {
+        return (brightness);
+    }
+}
+
+
+void TechnoClassExt::_Flashing_AI()
+{
+    int flash = FlashCount;
+    unsigned b = ((FlashCount / 2) % 2) == 1;
+    TechnoClassExtension* technoext = Extension::Fetch(this);
+
+    if (FlasherClass::Process() || technoext->IronCurtainTimer > 0) {
+        Mark(MARK_CHANGE);
+    }
+
+    if (RTTI == RTTI_BUILDING && (technoext->IronCurtainTimer > 0 || (flash && flash != FlashCount))) {
+        unsigned b2 = ((FlashCount / 2) % 2) == 1;
+        if (b != (unsigned char)b2) {
+            TacticalMap->Register_Dirty_Area(entry_118(), false);
+            ((BuildingClass*)this)->Update_Anim_Appearance();
+        }
+    }
+}
+
+/**
+ *  Makes the game redraw an object while it is flashing with the Iron Curtain effect.
+ *
+ *  Author: Rampastring
+ */
+DEFINE_HOOK(0x0062ECE3, _TechnoClass_AI_Iron_Curtain_Flash_Redraw_Patch, 0)
+{
+    GET(TechnoClassExt*, this_ptr, ESI);
+
+    this_ptr->_Flashing_AI();
+
+    return 0x0062ED7A;
+}
+
+
+/**
  *  Main function for patching the hooks.
  */
 void TechnoClassExtension_Hooks()
@@ -3217,4 +3306,5 @@ void TechnoClassExtension_Hooks()
     Patch_Jump(0x00638090, &TechnoClassExt::_Refund_Amount);
     Patch_Jump(0x0062D0F0, &TechnoClassExt::_Evaluate_Object);
     Patch_Jump(0x00638CA0, &TechnoClassExt::_Should_Self_Heal_Now);
+    Patch_Jump(0x00639C70, &TechnoClassExt::_Apparent_Brightness);
 }