+++ src/extensions/technotype/technotypeext.h
@@ -396,4 +396,9 @@ public:
      *  Should this unit not spawn wakes when it's cloaked? Usually useful for submarines.
      */
     bool IsHideWakeWhenCloaked;
+
+    /**
+     *  Specifies whether the AI should use the Iron Curtain to protect this object.
+     */
+    bool IronCurtainPriorityTarget;
 };