+++ docs/Whats-New.md
@@ -84,6 +84,7 @@ New:
 - Make `SOUND01.INI` load additively with `SOUND.INI`, reload sounds after loading side `MIX` files (by ZivDero)
 - Add "Building Does Not Exist" trigger event (by Rampastring)
 - Add "Create Building At" trigger action (by Rampastring)
+- Add the Iron Curtain logic from Red Alert 1 for map scripting and AI (by Rampastring)
 
 Vinifera fixes:
 - Fix unit placement in non-TS Client builds of Vinifera (by ZivDero)